cloud/holidays-get-holidays/requirements.txt
requests
google-cloud-firestore
google-cloud-error-reporting
